+++ packages/calcite-components/src/components/tree-item/tree-item.tsx
@@ -43,7 +43,7 @@ export class TreeItem extends LitElement implements InteractiveComponent {
 
   //#region Private Properties
 
-  private actionSlotWrapper = createRef<HTMLDivElement>();
+  private actionSlotWrapperRef = createRef<HTMLDivElement>();
 
   private childTree: Tree["el"];
 
@@ -129,6 +129,9 @@ export class TreeItem extends LitElement implements InteractiveComponent {
   /** Fires when the component's content area is expanded. */
   calciteTreeItemExpand = createEvent({ cancelable: false });
 
+  /** Fires when the component is selected or deselected. */
+  calciteTreeItemSelect = createEvent({ cancelable: false });
+
   //#endregion
 
   //#region Lifecycle
@@ -196,6 +199,11 @@ export class TreeItem extends LitElement implements InteractiveComponent {
         modifyCurrentSelection: true,
         updateItem: false,
       });
+    }
+
+    if (this.userChangedValue) {
+      this.calciteTreeItemSelect.emit();
+      this.userChangedValue = false;
     }
   }
 
@@ -291,7 +299,7 @@ export class TreeItem extends LitElement implements InteractiveComponent {
 
   private isActionEndEvent(event: Event): boolean {
     const composedPath = event.composedPath();
-    return composedPath.includes(this.actionSlotWrapper.value);
+    return composedPath.includes(this.actionSlotWrapperRef.value);
   }
 
   /**
@@ -489,7 +497,7 @@ export class TreeItem extends LitElement implements InteractiveComponent {
                 defaultSlotNode
               )}
             </div>
-            <div class={CSS.actionsEnd} hidden={!hasEndActions} ref={this.actionSlotWrapper}>
+            <div class={CSS.actionsEnd} hidden={!hasEndActions} ref={this.actionSlotWrapperRef}>
               {slotNode}
             </div>
           </div>