+++ packages/calcite-components-react/CHANGELOG.md
@@ -3,6 +3,34 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.0.0-next.3](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@4.0.0-next.2...@esri/calcite-components-react@4.0.0-next.3) (2025-09-24)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [4.0.0-next.2](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@4.0.0-next.1...@esri/calcite-components-react@4.0.0-next.2) (2025-09-24)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [4.0.0-next.1](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@4.0.0-next.0...@esri/calcite-components-react@4.0.0-next.1) (2025-09-24)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [4.0.0-next.0](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.3.1-next.3...@esri/calcite-components-react@4.0.0-next.0) (2025-09-24)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.3.1-next.3](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.3.1-next.2...@esri/calcite-components-react@3.3.1-next.3) (2025-09-24)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.3.1-next.2](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.3.1-next.1...@esri/calcite-components-react@3.3.1-next.2) (2025-09-23)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.3.1-next.1](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.3.1-next.0...@esri/calcite-components-react@3.3.1-next.1) (2025-09-19)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
 ## [3.3.1-next.0](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.3.0-next.116...@esri/calcite-components-react@3.3.1-next.0) (2025-09-19)
 
 **Note:** Version bump only for package @esri/calcite-components-react